MORE DAKKA!
Status:

Currently working on Dredd, which finally has the beginnings of a firing system. Going to start working on a movement system for Gump now, then go back to Dredd and work on a targetting system.

Once rudimentary versions of all three systems are in place, I'll move them to SoilentGreen for testing.

If it performs ok, I'll create three new test bots and beging developing better versions of these systems.